[YOLO] add pylint, move code to try avoid circular deps
