+++ custom_model_runner/datarobot_drum/drum/language_predictors/base_language_predictor.py
@@ -15,7 +15,8 @@ import pandas as pd
 
 from datarobot_drum.drum.adapters.cli.shared.drum_class_label_adapter import DrumClassLabelAdapter
 from datarobot_drum.drum.adapters.model_adapters.python_model_adapter import RawPredictResponse
-from datarobot_drum.drum.common import read_model_metadata_yaml, to_bool
+from datarobot_drum.drum.common import to_bool
+from datarobot_drum.drum.model_metadata import read_model_metadata_yaml
 from datarobot_drum.drum.enum import (
     LOGGER_NAME_PREFIX,
     ModelInfoKeys,
