+++ custom_model_runner/datarobot_drum/drum/push.py
@@ -12,8 +12,8 @@ import datarobot as dr_client
 
 from datarobot_drum.drum.common import (
     get_metadata,
-    validate_config_fields,
 )
+from datarobot_drum.drum.model_metadata import validate_config_fields
 from datarobot_drum.drum.enum import RunMode, TargetType, ModelMetadataKeys, ArgumentsOptions
 from datarobot_drum.drum.exceptions import DrumCommonException
