+++ custom_model_runner/datarobot_drum/runtime_parameters/runtime_parameters.py
@@ -6,29 +6,35 @@ Released under the terms of DataRobot Tool and Utility Agreement.
 """
 import json
 import os
-from collections import namedtuple, Counter
+from collections import namedtuple
 
 import trafaret as t
 import yaml
 
-from ..drum.enum import MODEL_CONFIG_FILENAME
-from ..drum.enum import ModelMetadataKeys
-from .exceptions import ErrorLoadingRuntimeParameter
-from .exceptions import InvalidEmptyYamlContent
-from .exceptions import InvalidInputFilePath
-from .exceptions import InvalidJsonException
-from .exceptions import InvalidRuntimeParam
-from .exceptions import InvalidYamlContent
-from .runtime_parameters_schema import (
+from datarobot_drum.drum.enum import MODEL_CONFIG_FILENAME
+from datarobot_drum.drum.enum import ModelMetadataKeys
+from datarobot_drum.runtime_parameters.exceptions import ErrorLoadingRuntimeParameter
+from datarobot_drum.runtime_parameters.exceptions import InvalidEmptyYamlContent
+from datarobot_drum.runtime_parameters.exceptions import InvalidInputFilePath
+from datarobot_drum.runtime_parameters.exceptions import InvalidJsonException
+from datarobot_drum.runtime_parameters.exceptions import InvalidRuntimeParam
+from datarobot_drum.runtime_parameters.exceptions import InvalidYamlContent
+from datarobot_drum.runtime_parameters.runtime_parameters_schema import (
     RuntimeParameterCredentialPayloadTrafaret,
     RuntimeParameterDeploymentPayloadTrafaret,
     RuntimeParameterBooleanPayloadTrafaret,
     RuntimeParameterNumericPayloadTrafaret,
 )
-from .runtime_parameters_schema import RuntimeParameterPayloadTrafaret
-from .runtime_parameters_schema import RuntimeParameterStringPayloadTrafaret
-from .runtime_parameters_schema import RuntimeParameterTypes
-from .runtime_parameters_schema import RuntimeParameterDefinitionTrafaret
+from datarobot_drum.runtime_parameters.runtime_parameters_schema import (
+    RuntimeParameterPayloadTrafaret,
+)
+from datarobot_drum.runtime_parameters.runtime_parameters_schema import (
+    RuntimeParameterStringPayloadTrafaret,
+)
+from datarobot_drum.runtime_parameters.runtime_parameters_schema import RuntimeParameterTypes
+from datarobot_drum.runtime_parameters.runtime_parameters_schema import (
+    RuntimeParameterDefinitionTrafaret,
+)
 
 
 class RuntimeParameters:
