+++ requirements_lint.txt
@@ -1,2 +1,3 @@
 black==23.10.1
+pylint==3.1.0
 click==8.1.7
